fixes and addition: read below
added testing jsons that aren't ignored
went back to letting huggingfaces control pinging
rename full-throttle-integration to email_sort_summary
added a email_categorization_screen just to demo it works on the phone (the responses are read in the debugging terminal)
