+++ assets/data/uncategorized_emails_10.json
@@ -1,0 +1,42 @@
+[
+    {
+        "Subject": "RE: NERC Statements on Impact of Security Threats on RTOs",
+        "Body": "I agree with Joe. The IOUs will point to NERC as an objective third party on these issues. "
+    },
+    {
+        "Subject": "RE: NERC Meeting Today",
+        "Body": "There was an all day meeting of the NERC/reliability legislation group today. I will provide a more detailed report but the group completed the process of reviewing the changes that some had suggested to shorten and streamline the NERC electric reliability organization legislation. Sarah and I asked a series of questions and made comments on our key issues and concerns. I want to give you a more complete report once I have gone back over the now final draft version. The timing being imposed by NERC is that they will circulate a clean version of the proposal tomorrow or Monday. They have asked for comments by next Thursday August 16th with an indication of whether each company/organization does or does not sign on to support it. They will then transmit the proposal and the endorsement letter to Congress and the Administration so they have it as Hill and Energy Dept. staff work on electricity drafting issues this month. I pointed out that EPSA is not due to meet internally with its members to discuss these issues until after the NERC deadline. That is not deterring NERC from moving forward with the above time frame."
+    },
+    {
+        "Subject": "http://www.stanford.edu/~wolak/",
+        "Body": "nan"
+    },
+    {
+        "Subject": "Re: Govt School Alumni",
+        "Body": "FYI -----------------"
+    },
+    {
+        "Subject": "Re: Data responses in SDG&E -- confidential treatment",
+        "Body": "PRIVILEGED AND CONFIDENTIAL Appropos of Agave's phone message. Steve was in this loop. The discount reports are available to anyone who wants them under the Open Records Act. TK you indicated that Agave thinks the rates were published somewhere by FERC -- I don't know whether that's true but the bottom line is that FERC has the information in its files and could publish it any time at wants we have no control over that. Let me know if you want me to talk with Agave's lawyers. -----------------"
+    },
+    {
+        "Subject": "CONFIDENTIAL - Residential in CA",
+        "Body": "In the meeting today no decision was made about what to do with Enron's 16000 residential customers. Each of the contracts gives a basic 30 day out right to Enron. That being said I think that we have a short window to push for DA before any public action impacts us in Sacramento. I realize that the ultimate action (which I think is inevitable) makes it harder for our advocacy on DA but real $ are flowing out of the company. EES will give us notice when a decision is reached. Thanks Jim"
+    },
+    {
+        "Subject": "Thursday Meeting on Winter Gas Prices",
+        "Body": "calendar -----------------"
+    },
+    {
+        "Subject": "Re: Privileged and Confidential communication to my attorneys",
+        "Body": "Felicia This is a merchant asset which we currently own 100% but plan to sell down in the coming months. It absolutely has separate HR functions from Enron -- they do not participate in Enron benefits and we are working hard to prevent any claims that we could be a joint employer with it. Thoughts? Michelle Enron North America Corp. From: Felecia Acevedo @ ENRON 11/02/2000 08:56 AM To: Michelle Cash/HOU/ECT@ECT cc: Subject: Re: Privileged and Confidential communication to my attorneys Michelle who is GSP? Are they a legal entity? Are they on our payroll? If they are we filed EEO-1's for them. Below is the criteria to file an EEO-1 report: Standard Form 100 must be filed by - A. All private employers who are: (1) subject to Title VII of the Civil Rights Act of 1964 (as amended by the Equal Employment Opportunity Act of 1972) with 100 or more employees EXCLUDING State and local governments primary and secondary school systems institutions of higher education Indian tribes and tax-exempt private membership clubs other than labor organizations OR (2) subject to Title VII who have fewer than 100 employees if the company is owned or affiliated with another company or there is centralized ownership control or management (such as central control of personnel policies and labor relations) so that the group legally constitutes a single enterprise and the entire enterprise employs a total of 100 or more employees. B. All federal contractors (private employers) who: (1) are not exempt as provided for by 41 CFR 60-1.5 (2) have 50 or more employees and (a) are prime contractors or first-tier subcontractors and have a contract subcontract or purchase order amounting to $50000 or more or (b) serve as a depository of Government funds in any amount or (c) is a financial institution which is an issuing and paying agent for U.S. Savings Bonds and Notes. All multi-establishment employers i.e. employers doing business at more than one establishment must file: (1) a report covering the principal or headquarters office (2) a separate report for each establishment employing 50 or more persons (3) a consolidated report that MUST include ALL employees by race sex and job category in establishments with 50 or more employees as well as establishments with fewer than 50 employees and (4) a list showing the name address total employment and major activity for each establishment employ-ing fewer than 50 persons must accompany the consolidated report. All forms for a multi-establishment company must be collected by the headquarters office for its establishments or by the parent corporation for its subsidiary holdings and submitted in one package. For the purposes of this report the term parent corporation refers to any corporation which owns all or the majority stock of another corporation so that the latter stands in the relation to it of a subsidiary. The OFCCP has always held that subsidiary companies that do not have direct govt contracts must comply with AAP regulations if the parent company or other subsidiary companies benefit from government contracts. Their line of thought is that the subsidiary without government contracts benefits indirectly from the contract as well. If GSP is completely free standing and they do not participate in Enron's benefits abide by our policies or procedures  participate in our bonus process that may partially or wholly be funded by Enron Corp.  etc. they may have some type of argement not to comply but I doubt it. If we own them over 51% I don't see a way they could not file EEO-1's and have an AAP. We do all the AAP's in Corporate so if this is a freestanding Business Unit I need to know. Thanks! Felecia Michelle Cash@ECT 11/01/2000 07:49 PM To: Felecia Acevedo/Corp/Enron@Enron cc: Subject: Privileged and Confidential communication to my attorneys Felicia Does this sound right to you? Just wanted to confirm. Thanks. Michelle -----------------"
+    },
+    {
+        "Subject": "Rice Letter",
+        "Body": "Anita Very minor changes. Thanks. Vince"
+    },
+    {
+        "Subject": "Enron Wind Policy Meeting",
+        "Body": "-----------------"
+    }
+]
